pyproject.toml
[project]
name = "agent11p1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.19.0",
    "google-genai>=1.52.0",
    "python-dotenv>=1.2.1",
]
